+++ moddb/pages/file.py
@@ -1,6 +1,7 @@
 import datetime
 import re
 import sys
+from typing import BinaryIO, List
 
 import bs4
 import requests
@@ -151,20 +152,22 @@ class File(BaseMetaClass):
     def __repr__(self):
         return f"<{self.__class__.__name__} name={self.name} type={self.category.name}>"
 
-    def save(self, file_obj, *, mirror=None):
+    def save(self, file_obj: BinaryIO, *, mirror=None, chunk_size: int = 10_000_000):
         """Save the file to an object. This functions makes
         two requests. If you pass a valid mirror it will
         make only one request.
 
         Parameters
         -----------
-        file_obj : typing.BinaryIO
+        file_obj : BinaryIO
             The file obj to save the file to. The binary data
             will be streamed to that object.
         mirror : Optional[Mirror]
             An optional mirror object to download the
             file from a specific moddb mirror
-
+        chunk_size: int
+            The size of the chunks to stream the response 
+            back in. 10MB by default
         """
         if mirror is None:
             download = get_page(f"{BASE_URL}/downloads/start/{self.id}")
@@ -172,14 +175,14 @@ class File(BaseMetaClass):
         else:
             url = mirror._url
 
-        SESSION = sys.modules["moddb"].SESSION
+        SESSION: requests.Session = sys.modules["moddb"].SESSION
         prepped = prepare_request(requests.Request("GET", join(url)), SESSION)
         with SESSION.send(prepped, stream=True) as r:
-            raise_for_status(r)
-            for chunk in r.iter_content(chunk_size=8192):
+            r.raise_for_status()
+            for chunk in r.iter_content(chunk_size=chunk_size):
                 file_obj.write(chunk)
 
-    def get_mirrors(self):
+    def get_mirrors(self) -> List[Mirror]:
         """Get all the mirrors from which a file can be downloaded. This
         can then be passed to File.save to download from a specific mirror.
 
@@ -361,20 +364,23 @@ class Media(BaseMetaClass):
     def __repr__(self):
         return f"<Media name={self.name} type={self.category.name}>"
 
-    def save(self, file_obj):
+    def save(self, file_obj: BinaryIO, *, chunk_size: int = 10_000_000):
         """Save the media to an object.
 
         Parameters
         -----------
-        file_obj : typing.BinaryIO
+        file_obj : BinaryIO
             The file obj to save the file to. The binary data
             will be streamed to that object.
+        chunk_size: int
+            The size of the chunks to stream the response 
+            back in. 10MB by default
 
         """
-        SESSION = sys.modules["moddb"].SESSION
+        SESSION: requests.Session = sys.modules["moddb"].SESSION
         prepped = prepare_request(requests.Request("GET", self.fileurl), SESSION)
 
         with SESSION.send(prepped, stream=True) as r:
-            raise_for_status(r)
-            for chunk in r.iter_content(chunk_size=8192):
+            r.raise_for_status()
+            for chunk in r.iter_content(chunk_size=chunk_size):
                 file_obj.write(chunk)